+++ experimental/packages/exporter-logs-otlp-grpc/test/logsHelper.ts
@@ -88,7 +88,7 @@ export function ensureExportedLogRecordIsCorrect(logRecord: ILogRecord) {
   ensureExportedAttributesAreCorrect(logRecord.attributes);
   assert.strictEqual(
     logRecord.timeUnixNano,
-    '1680253513123241728',
+    '1680253513123241635',
     'timeUnixNano is wrong'
   );
   assert.strictEqual(