+++ experimental/packages/exporter-trace-otlp-http/test/node/CollectorTraceExporter.test.ts
@@ -316,7 +316,6 @@ describe('OTLPTraceExporter - node with json over http', () => {
 
       fakeRequest.on('end', () => {
         const responseBody = buff.toString();
-
         const json = JSON.parse(responseBody) as IExportTraceServiceRequest;
         const span1 = json.resourceSpans?.[0].scopeSpans?.[0].spans?.[0];
         assert.ok(typeof span1 !== 'undefined', "span doesn't exist");