+++ experimental/packages/opentelemetry-exporter-metrics-otlp-grpc/test/metricsHelper.ts
@@ -17,6 +17,7 @@
 import {
   Counter,
   Histogram,
+  HrTime,
   ObservableGauge,
   ObservableResult,
   ValueType,
@@ -31,7 +32,13 @@ import {
   MetricReader,
   View,
 } from '@opentelemetry/sdk-metrics';
-import { IKeyValue, IMetric, IResource } from '@opentelemetry/otlp-transformer';
+import {
+  hrTimeToFixed64Nanos,
+  IKeyValue,
+  IMetric,
+  IResource,
+  UnsignedLong,
+} from '@opentelemetry/otlp-transformer';
 
 class TestMetricReader extends MetricReader {
   protected onForceFlush(): Promise<void> {
@@ -125,92 +132,99 @@ export function ensureExportedAttributesAreCorrect(attributes: IKeyValue[]) {
 
 export function ensureExportedCounterIsCorrect(
   metric: IMetric,
-  time?: number,
-  startTime?: number
+  time: HrTime,
+  startTime: HrTime
 ) {
-  assert.deepStrictEqual(metric, {
-    name: 'int-counter',
-    description: 'sample counter description',
-    unit: '',
-    data: 'sum',
-    sum: {
-      dataPoints: [
-        {
-          attributes: [],
-          exemplars: [],
-          value: 'asInt',
-          asInt: '1',
-          flags: 0,
-          startTimeUnixNano: String(startTime),
-          timeUnixNano: String(time),
-        },
-      ],
-      isMonotonic: true,
-      aggregationTemporality: 'AGGREGATION_TEMPORALITY_CUMULATIVE',
-    },
-  });
+  assert.strictEqual(metric.name, 'int-counter');
+  assert.strictEqual(metric.description, 'sample counter description');
+  assert.strictEqual(metric.unit, '');
+  assert.strictEqual(metric.sum?.dataPoints.length, 1);
+  assert.strictEqual(
+    metric.sum?.aggregationTemporality,
+    'AGGREGATION_TEMPORALITY_CUMULATIVE'
+  );
+  assert.strictEqual(metric.sum?.isMonotonic, true);
+
+  const [dp] = metric.sum.dataPoints;
+
+  assert.deepStrictEqual(dp.attributes, []);
+  assert.deepStrictEqual(dp.exemplars, []);
+  assert.strictEqual(dp.asInt, '1');
+  assert.strictEqual(dp.flags, 0);
+
+  assert.deepStrictEqual(
+    UnsignedLong.fromString(dp.startTimeUnixNano as string),
+    hrTimeToFixed64Nanos(startTime)
+  );
+  assert.deepStrictEqual(
+    UnsignedLong.fromString(dp.timeUnixNano as string),
+    hrTimeToFixed64Nanos(time)
+  );
 }
 
 export function ensureExportedObservableGaugeIsCorrect(
   metric: IMetric,
-  time?: number,
-  startTime?: number
+  time: HrTime,
+  startTime: HrTime
 ) {
-  assert.deepStrictEqual(metric, {
-    name: 'double-observable-gauge',
-    description: 'sample observable gauge description',
-    unit: '',
-    data: 'gauge',
-    gauge: {
-      dataPoints: [
-        {
-          attributes: [],
-          exemplars: [],
-          value: 'asDouble',
-          asDouble: 6,
-          flags: 0,
-          startTimeUnixNano: String(startTime),
-          timeUnixNano: String(time),
-        },
-      ],
-    },
-  });
+  assert.strictEqual(metric.name, 'double-observable-gauge');
+  assert.strictEqual(metric.description, 'sample observable gauge description');
+  assert.strictEqual(metric.unit, '');
+  assert.strictEqual(metric.gauge?.dataPoints.length, 1);
+
+  const [dp] = metric.gauge.dataPoints;
+
+  assert.deepStrictEqual(dp.attributes, []);
+  assert.deepStrictEqual(dp.exemplars, []);
+  assert.strictEqual(dp.asDouble, 6);
+  assert.strictEqual(dp.flags, 0);
+
+  assert.deepStrictEqual(
+    UnsignedLong.fromString(dp.startTimeUnixNano as string),
+    hrTimeToFixed64Nanos(startTime)
+  );
+  assert.deepStrictEqual(
+    UnsignedLong.fromString(dp.timeUnixNano as string),
+    hrTimeToFixed64Nanos(time)
+  );
 }
 
 export function ensureExportedHistogramIsCorrect(
   metric: IMetric,
-  time?: number,
-  startTime?: number,
+  time: HrTime,
+  startTime: HrTime,
   explicitBounds: number[] = [Infinity],
   bucketCounts: string[] = ['2', '0']
 ) {
-  assert.deepStrictEqual(metric, {
-    name: 'int-histogram',
-    description: 'sample histogram description',
-    unit: '',
-    data: 'histogram',
-    histogram: {
-      dataPoints: [
-        {
-          attributes: [],
-          exemplars: [],
-          flags: 0,
-          _sum: 'sum',
-          _min: 'min',
-          _max: 'max',
-          sum: 21,
-          count: '2',
-          min: 7,
-          max: 14,
-          startTimeUnixNano: String(startTime),
-          timeUnixNano: String(time),
-          bucketCounts,
-          explicitBounds,
-        },
-      ],
-      aggregationTemporality: 'AGGREGATION_TEMPORALITY_CUMULATIVE',
-    },
-  });
+  assert.strictEqual(metric.name, 'int-histogram');
+  assert.strictEqual(metric.description, 'sample histogram description');
+  assert.strictEqual(metric.unit, '');
+  assert.strictEqual(metric.histogram?.dataPoints.length, 1);
+  assert.strictEqual(
+    metric.histogram?.aggregationTemporality,
+    'AGGREGATION_TEMPORALITY_CUMULATIVE'
+  );
+
+  const [dp] = metric.histogram.dataPoints;
+
+  assert.deepStrictEqual(dp.attributes, []);
+  assert.deepStrictEqual(dp.exemplars, []);
+  assert.strictEqual(dp.flags, 0);
+  assert.strictEqual(dp.sum, 21);
+  assert.strictEqual(dp.count, '2');
+  assert.strictEqual(dp.min, 7);
+  assert.strictEqual(dp.max, 14);
+
+  assert.deepStrictEqual(
+    UnsignedLong.fromString(dp.startTimeUnixNano as string),
+    hrTimeToFixed64Nanos(startTime)
+  );
+  assert.deepStrictEqual(
+    UnsignedLong.fromString(dp.timeUnixNano as string),
+    hrTimeToFixed64Nanos(time)
+  );
+  assert.deepStrictEqual(dp.bucketCounts, bucketCounts);
+  assert.deepStrictEqual(dp.explicitBounds, explicitBounds);
 }
 
 export function ensureResourceIsCorrect(resource: IResource) {