+++ experimental/CHANGELOG.md
@@ -56,6 +56,7 @@ All notable changes to experimental packages in this project will be documented 
 
 ### :bug: (Bug Fix)
 
+* fix(otlp-transformer): Avoid precision loss when converting from HrTime to unix nanoseconds. [#4062](https://github.com/open-telemetry/opentelemetry-js/pull/4062)
 * fix(exporter-logs-otlp-http): add @opentelemetry/api-logs as dependency
 
 ## 0.41.2